Minimax Rev2
Revision 2 of the Minimax algorithm within the game. The algorithm doesn't yet run for board sizes over 3, though it appears that it would be dynamic and theoretically work for these instances. To the best of my knowledge the issue at hand for sizes above 3 is the algorithm complexity being boardsize^2 and taking extremely long to process. Hopefully the next revision of the algorithm will fix this using a method such as Alpha-Beta pruning and maybe hash tables.