Minimax Rev 3
Alpha-Beta pruning implemented and greatly decreases the complexity and processing time, though due to the pure size of numbers such as 5^2, the only functional board sizes are 3 and 4, where functional means a reasonable amount of time before a decision is made and the player can have their turn. The use of Zobrist Hashing in a Transposition table is most-likely the most effective way to assist in decreasing this wait for results.