Monte Carlo Rev 1
Revision 1 of the AI using the Monte Carlo simulation/algorithm. Not currently functional though the project will build and run.
Did a minor cleanup on the minimax code as well as some small functional changes. After research and testing I decided against using Transposition Table at this point and instead am testing methods such as Iterative Deepening (IDDFS), Negascout/PVS, MTD(f), and Aspiration Windows as the next steps for optimising the minimax algorithm, though most will probably not be in the final product in order to not overcomplicate it.
A draft version of Principle Variation Search (PVS) is in the files now, though not functional. It's under the name Negascout due to the fact that it will eventually be modified to be more like that variant.