Repaired Monte Carlo & Changed to Cells
Fixed bugs in the Monte Carlo player algorithm allowing for optimized use within both AI simulations and human games. Converted the vector systems from using indexing to using a Cell struct. This was done for both simplicity and optimization , with hopes that overall, the use of references to Cell items will be better than the process of calculating coordinate indices.